Add the option to show the users who have bought tickets on an event
This helps with the confusion when for instance with a symposium you want to create hype but users get confused by only getting a ticket or only signing up. The idea is that you then will not make an event for that anymore.
Fixes #1900 and #1820 as you can add two free tickets.