Prevent seeder crashes when event has bad dates.
This commit makes it so when the signup start is after the start of the
activity, the seeder sends a warning and sets the signup date to the
start date of the activity.

This behaviour is not fully correct and the activity should be corrected
on the live site, thats why it sends a warning.